ejer2 sem3

+++ SEMANA-1_C#/Program.cs
@@ -25,14 +25,14 @@ namespace SEMANA_1_C_
             Console.ReadKey ();
 
         }
-        static void ejer2() 
+        static void ejer2()
         {
             Console.Write("ingrese numero x: ");
-            int x = int.Parse( Console.ReadLine());
+            int x = int.Parse(Console.ReadLine());
             Console.Write("ingrese numero y: ");
-            int y = int.Parse( Console.ReadLine());
+            int y = int.Parse(Console.ReadLine());
 
-            double resu = (x / y);
+            double resu = ((double) x / (double) y);
 
             Console.WriteLine("SUMA: " + (x + y));
             Console.WriteLine("RESTA: " + (x - y));
